+++ Changelog.md
@@ -3,6 +3,9 @@
 ## V 1.3.0 : XXX Update - XX/XX/19
 
     - loggers : Create logging System
+    - Entity : Can't have the same type of component two times
+    
+    - Crash Fix : Crash when show id of Entity Texts
 
 ## V 1.2.0 : Property Update - 09/06/19